requirements.txt
mouse==0.7.1
Pillow==9.3.0
pywin32==304
torch==1.13.0
torchvision==0.14.0
